Hostapd: add support to mtlk (Lantiq wireless)
Original patches have been taken from UGW 7.1.1 and the names have been kept.

It seems that UGW was updated up to v2.2.

Path 715 is a syntax adaptation to libnl2. In this regards, the deletion of "mgmt->u.probe_req.variable" is a guess. I couldn't find exactly the meaning of it in other examples so this might be a source of bugs.

iSTATUS:	It may work. So far it launches a message "initialization failed" that might be due to the driver, not to hostapd

Signed-off-by: Jorge Amorós-Argos <joramar76@gmail.com>